Make GeneratePopoverBody as independent component
