+++ public/components/incontext_insight/index.tsx
@@ -26,15 +26,12 @@ import {
   EuiButtonIcon,
 } from '@elastic/eui';
 import React, { Children, isValidElement, useEffect, useRef, useState } from 'react';
-import { integer } from '@opensearch-project/opensearch/api/types';
 import { IncontextInsight as IncontextInsightInput } from '../../types';
 import { getNotifications, IncontextInsightRegistry } from '../../services';
 // TODO: Replace with getChrome().logos.Chat.url
 import chatIcon from '../../assets/chat.svg';
-import { ASSISTANT_API } from '../../../common/constants/llm';
 import { HttpSetup } from '../../../../../src/core/public';
-import { Interaction } from '../../../common/types/chat_saved_object_attributes';
-import { getAssistantRole } from '../../../server/services/chat/olly_chat_service';
+import { GeneratePopoverBody } from './generate_popover_body';
 
 export interface IncontextInsightProps {
   children?: React.ReactNode;
@@ -52,10 +49,6 @@ export const IncontextInsight = ({
 }: IncontextInsightProps) => {
   const anchor = useRef<HTMLDivElement>(null);
   const [isVisible, setIsVisible] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-  const [isLlmResponded, setIsLlmResponded] = useState(false);
-  const [summary, setSummary] = useState('');
-  const [conversationId, setConversationId] = useState('');
 
   useEffect(() => {
     // TODO: use animation when not using display: none
@@ -157,70 +150,6 @@ export const IncontextInsight = ({
     }
   };
 
-  const onChatContinuation = (incontextInsight: IncontextInsightInput) => {
-    setIsVisible(false);
-    registry?.continueInChat(incontextInsight, conversationId);
-    if (anchor.current) {
-      const incontextInsightAnchorButtonClassList = anchor.current.parentElement?.querySelector(
-        '.incontextInsightAnchorButton'
-      )?.classList;
-      incontextInsightAnchorButtonClassList?.remove('incontextInsightHoverEffectUnderline');
-    }
-  };
-
-  const onGenerateSummary = (
-    incontextInsight: IncontextInsightInput,
-    summarizationQuestion: string
-  ) => {
-    setIsLoading(true);
-    const summarize = async () => {
-      const contextContent = incontextInsight.contextProvider
-        ? await incontextInsight.contextProvider()
-        : '';
-
-      await httpSetup
-        ?.post(ASSISTANT_API.SEND_MESSAGE, {
-          body: JSON.stringify({
-            messages: [],
-            input: {
-              type: 'input',
-              content: summarizationQuestion,
-              contentType: 'text',
-              context: { content: contextContent, dataSourceId: incontextInsight.datasourceId },
-              promptPrefix: getAssistantRole(incontextInsight.key),
-            },
-          }),
-        })
-        .then((response) => {
-          response.interactions.map(
-            (interaction: Interaction, index: integer, array: Interaction[]) => {
-              if (index === array.length - 1) {
-                setConversationId(interaction.conversation_id);
-              }
-            }
-          );
-
-          response.messages.map((message: { type: string; content: string }) => {
-            if (message.type === 'output') {
-              setSummary(message.content);
-            }
-          });
-        })
-        .catch((error) => {
-          toasts.addDanger(
-            i18n.translate('assistantDashboards.incontextInsight.generateSummaryError', {
-              defaultMessage: 'Generate summary error',
-            })
-          );
-          setIsLoading(false);
-        });
-
-      return;
-    };
-
-    return summarize();
-  };
-
   const SuggestionsPopoverFooter: React.FC<{ incontextInsight: IncontextInsightInput }> = ({
     incontextInsight,
   }) => (
@@ -258,86 +187,13 @@ export const IncontextInsight = ({
     </EuiPopoverFooter>
   );
 
-  const GeneratePopoverBody: React.FC<{ incontextInsight: IncontextInsightInput }> = ({
-    incontextInsight,
-  }) => {
-    if (!isLoading)
-      return (
-        <EuiButton
-          onClick={async () => {
-            await onGenerateSummary(
-              incontextInsight,
-              incontextInsight.suggestions && incontextInsight.suggestions.length > 0
-                ? incontextInsight.suggestions[0]
-                : 'Please summarize the input'
-            );
-            setIsLlmResponded(true);
-          }}
-        >
-          {i18n.translate('assistantDashboards.incontextInsight.generateSummary', {
-            defaultMessage: 'Generate summary',
-          })}
-        </EuiButton>
-      );
-    if (isLoading && !isLlmResponded)
-      return (
-        <EuiButton isLoading={isLoading}>
-          {i18n.translate('assistantDashboards.incontextInsight.generatingSummary', {
-            defaultMessage: 'Generating summary...',
-          })}
-        </EuiButton>
-      );
-    if (isLoading && isLlmResponded)
-      return (
-        <>
-          <SummaryPopoverBody incontextInsight={incontextInsight} />
-          <EuiSpacer size={'xs'} />
-          <EuiPanel
-            hasShadow={false}
-            hasBorder={false}
-            element="div"
-            onClick={() => {
-              onChatContinuation(incontextInsight);
-            }}
-            grow={false}
-            paddingSize="none"
-            style={{ width: '120px', float: 'right' }}
-          >
-            <EuiFlexGroup gutterSize="none" style={{ marginTop: 5 }}>
-              <EuiFlexItem grow={false}>
-                <EuiIcon type={'chatRight'} style={{ marginRight: 5 }} />
-              </EuiFlexItem>
-              <EuiFlexItem>
-                <EuiText size="xs">
-                  {i18n.translate('assistantDashboards.incontextInsight.continueInChat', {
-                    defaultMessage: 'Continue in chat',
-                  })}
-                </EuiText>
-              </EuiFlexItem>
-            </EuiFlexGroup>
-          </EuiPanel>
-        </>
-      );
-  };
-
   const SummaryPopoverBody: React.FC<{ incontextInsight: IncontextInsightInput }> = ({
     incontextInsight,
-  }) => {
-    // When there are multiple component objects with different summaries, use summary state as body
-    if (summary !== '') {
-      return (
-        <EuiPanel paddingSize="s" hasBorder hasShadow={false} color="plain">
-          <EuiText size="s">{summary}</EuiText>
-        </EuiPanel>
-      );
-    } else {
-      return (
-        <EuiPanel paddingSize="s" hasBorder hasShadow={false} color="plain">
-          <EuiText size="s">{incontextInsight.summary}</EuiText>
-        </EuiPanel>
-      );
-    }
-  };
+  }) => (
+    <EuiPanel paddingSize="s" hasBorder hasShadow={false} color="plain">
+      <EuiText size="s">{incontextInsight.summary}</EuiText>
+    </EuiPanel>
+  );
 
   const SummaryWithSuggestionsPopoverBody: React.FC<{
     incontextInsight: IncontextInsightInput;
@@ -348,42 +204,45 @@ export const IncontextInsight = ({
     </>
   );
 
-  const [userQuestion, setUserQuestion] = useState('');
   const ChatPopoverBody: React.FC<{ incontextInsight: IncontextInsightInput }> = ({
     incontextInsight,
-  }) => (
-    <EuiFlexGroup gutterSize="xs">
-      <EuiFlexItem grow={6}>
-        <EuiFormRow>
-          <EuiFieldText
-            placeholder="Ask a question"
-            value={userQuestion}
-            autoFocus
-            onChange={(e) => setUserQuestion(e.target.value)}
-            onKeyDown={(e) => {
-              if (e.key === 'Enter') {
-                onSubmitClick(incontextInsight, userQuestion);
-                setUserQuestion('');
-              }
+  }) => {
+    const [userQuestion, setUserQuestion] = useState('');
+
+    return (
+      <EuiFlexGroup gutterSize="xs">
+        <EuiFlexItem grow={6}>
+          <EuiFormRow>
+            <EuiFieldText
+              placeholder="Ask a question"
+              value={userQuestion}
+              autoFocus
+              onChange={(e) => setUserQuestion(e.target.value)}
+              onKeyDown={(e) => {
+                if (e.key === 'Enter') {
+                  onSubmitClick(incontextInsight, userQuestion);
+                  setUserQuestion('');
+                }
+              }}
+            />
+          </EuiFormRow>
+        </EuiFlexItem>
+        <EuiFlexItem grow={false}>
+          <EuiButton
+            fill
+            iconType="returnKey"
+            iconSide="right"
+            onClick={() => {
+              onSubmitClick(incontextInsight, userQuestion);
+              setUserQuestion('');
             }}
-          />
-        </EuiFormRow>
-      </EuiFlexItem>
-      <EuiFlexItem grow={false}>
-        <EuiButton
-          fill
-          iconType="returnKey"
-          iconSide="right"
-          onClick={() => {
-            onSubmitClick(incontextInsight, userQuestion);
-            setUserQuestion('');
-          }}
-        >
-          Go
-        </EuiButton>
-      </EuiFlexItem>
-    </EuiFlexGroup>
-  );
+          >
+            Go
+          </EuiButton>
+        </EuiFlexItem>
+      </EuiFlexGroup>
+    );
+  };
 
   const ChatWithSuggestionsPopoverBody: React.FC<{ incontextInsight: IncontextInsightInput }> = ({
     incontextInsight,
@@ -425,7 +284,14 @@ export const IncontextInsight = ({
         case 'suggestions':
           return <SuggestionsPopoverFooter incontextInsight={input} />;
         case 'generate':
-          return <GeneratePopoverBody incontextInsight={input} />;
+          return (
+            <GeneratePopoverBody
+              incontextInsight={input}
+              httpSetup={httpSetup}
+              registry={registry}
+              closePopover={closePopover}
+            />
+          );
         case 'summary':
           return <SummaryPopoverBody incontextInsight={input} />;
         case 'summaryWithSuggestions':
